refactor: factor module-to-script out of test-gimix-proposal
